+++ assets/js/src/public/templates/banners.js
@@ -1,5 +1,5 @@
-export const bannerContent = content => (
-	`<div class="bc-banner">${content}</div>`
+export const bannerContent = (content, label) => (
+	`<aside class="bc-banner" aria-label="${label}">${content}</aside>`
 );
 
 export const bannerWrapper = (styles = null, banners = null) => (