+++ src/BigCommerce/Webhooks/Webhook_Versioning.php
@@ -4,6 +4,9 @@ namespace BigCommerce\Webhooks;
 
 
 use BigCommerce\Settings\Sections\Import;
+use BigCommerce\Webhooks\Customer\Customer_Create_Webhook;
+use BigCommerce\Webhooks\Customer\Customer_Delete_Webhook;
+use BigCommerce\Webhooks\Customer\Customer_Update_Webhook;
 
 /**
  * Class Webhook_Versioning
@@ -37,7 +40,7 @@ class Webhook_Versioning {
 	 * @action bigcommerce/settings/webhoooks_updated
 	 */
 	public function maybe_update_webhooks() {
-		if ( ! get_option( Import::ENABLE_WEBHOOKS , 1 ) ) {
+		if ( ! $this->product_webhooks_enabled() && ! $this->customer_webhooks_enabled() ) {
 			return;
 		}
 
@@ -48,11 +51,39 @@ class Webhook_Versioning {
 		}
 	}
 
+	private function product_webhooks_enabled() {
+		return get_option( Import::ENABLE_PRODUCTS_WEBHOOKS , 1 );
+	}
+
+	private function customer_webhooks_enabled() {
+		return get_option( Import::ENABLE_CUSTOMER_WEBHOOKS , 1 );
+	}
+
 	/**
 	 * @return void Set new routes whenever any of the route list element gets updated
 	 */
 	private function update_webhooks() {
-		foreach ( $this->hooks as $hook ) {
+		$customer_webhooks         = [
+			Customer_Create_Webhook::NAME,
+			Customer_Update_Webhook::NAME,
+			Customer_Delete_Webhook::NAME,
+		];
+		$customer_webhooks_enabled = $this->customer_webhooks_enabled();
+		$product_webhooks_enabled  = $this->product_webhooks_enabled();
+
+		foreach ( $this->hooks as $key => $hook ) {
+			/**
+			 * Check if only product or customer webhooks are enabled
+			 * We don't need to update all webhooks
+			 */
+			$is_customer_webhook       = in_array( $key, $customer_webhooks );
+			$should_skip_customer_hook = ! $customer_webhooks_enabled && $is_customer_webhook;
+			$should_skip_product_hook  = ! $product_webhooks_enabled && ! $is_customer_webhook;
+
+			if ( $should_skip_customer_hook || $should_skip_product_hook ) {
+				continue;
+			}
+
 			$hook->update();
 		}
 	}